+++ src/apps/cc/CMakeLists.txt
@@ -19,9 +19,9 @@ set(
     datamodel/ddl.cpp
     detail/sqlite.cpp
     detector/arrival.cpp
-    detector/detector.cpp
+    detector/detector_impl.cpp
     detector/detectorbuilder.cpp
-    detector/detectorwaveformprocessor.cpp
+    detector/detector.cpp
     detector/linker/association.cpp
     detector/linker/pot.cpp
     detector/linker/strategy.cpp
